# Memoizes and caches provider settings to improve performance