#### :tophat: What? Why?
Improves performance of pages including omniauth links by memoizing their configuration and caching the related partials.

Main issue here is that the method `omniauth_provider_settings`, which does the parsing and decrypting of the omniauth provider settings, is called 5 times for each provider in a single request.

Quick debug log before the optimization, having 3 omniauth providers enabled, taking almost 1 second to render the partial:

```
  Rendered /decidim/decidim-core/app/views/layouts/decidim/_wrapper.html.erb (693.0ms)
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
  Rendered /decidim/decidim-core/app/views/decidim/devise/shared/_omniauth_buttons_mini.html.erb (943.9ms)
```

Just by memoizing the result, the process is run once from each provider and the rendering time for the partial is reduced to rougly 200ms:

```
  Rendered /decidim/decidim-core/app/views/layouts/decidim/_wrapper.html.erb (651.0ms)
In the omniauth_provider_settings method for twitter
In the omniauth_provider_settings method for facebook
In the omniauth_provider_settings method for google_oauth2
  Rendered /decidim/decidim-core/app/views/decidim/devise/shared/_omniauth_buttons_mini.html.erb (195.8ms)
```
 
#### :pushpin: Related Issues
- Related to #6141