@mrcasals 2020-06-23
Yay! Well, that's a start, glad to see it actually worked! 

There's room for improvement using the Rails cache though! 😄