@ace 2020-06-25
I wanted to go one step at a time, but, given the popular demand for caching (thanks @mrcasals and @armandfardeau 😄) I've added caching to the omniauth partials.