@mrcasals 2020-06-25
@armandfardeau it'd be awesome if you could try this PR and run performance tests to see how this improves it! 😄